pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "control-view-sidecar"
version = "0.1.0"
description = "Control View sidecar for PX4 supervisory control over ROS 2 and MCP."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastmcp>=2.11.0",
  "pydantic>=2.10.0",
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.12.0",
]

[project.scripts]
control-view-sidecar = "control_view.app:main"
control-view-raw-mcp = "control_view.raw_app:main"
control-view-observer = "control_view.observer.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
